refactor: 가독성을 위해 분리된 모듈 하나로 합침
input/makeNewCars 모듈을 getCarNames에 합침
